+++ lib/active_record/connection_adapters/sqlserver/schema_statements.rb
@@ -301,7 +301,10 @@ module ActiveRecord
             columns.NUMERIC_SCALE AS numeric_scale,
             columns.NUMERIC_PRECISION AS numeric_precision,
             columns.DATETIME_PRECISION AS datetime_precision,
-            columns.COLLATION_NAME AS collation,
+            CASE
+              WHEN columns.COLLATION_NAME <> DATABASEPROPERTYEX(DB_NAME(), 'Collation') THEN columns.COLLATION_NAME
+              ELSE NULL
+            END AS [collation],
             columns.ordinal_position,
             CASE
               WHEN columns.DATA_TYPE IN ('nchar','nvarchar','char','varchar') THEN columns.CHARACTER_MAXIMUM_LENGTH